+++ common/src/main/java/dev/mrsnowy/teleport_commands/commands/home.java
@@ -4,8 +4,14 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.datafixers.util.Pair;
 import dev.mrsnowy.teleport_commands.TeleportCommands;
 import dev.mrsnowy.teleport_commands.storage.StorageManager;
+import dev.mrsnowy.teleport_commands.common.NamedLocation;
+import dev.mrsnowy.teleport_commands.common.Player;
 import dev.mrsnowy.teleport_commands.suggestions.HomeSuggestionProvider;
+
+import java.util.Locale;
 import java.util.Objects;
+import java.util.Optional;
+
 import net.minecraft.ChatFormatting;
 import net.minecraft.commands.Commands;
 import net.minecraft.core.BlockPos;
@@ -15,9 +21,9 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.phys.Vec3;
 
+import static dev.mrsnowy.teleport_commands.storage.StorageManager.STORAGE;
 import static dev.mrsnowy.teleport_commands.utils.tools.getTranslatedText;
 import static net.minecraft.commands.Commands.argument;
-import static dev.mrsnowy.teleport_commands.storage.StorageManager.GetPlayerStorage;
 import static dev.mrsnowy.teleport_commands.storage.StorageManager.StorageSaver;
 import static dev.mrsnowy.teleport_commands.utils.tools.Teleporter;
 
@@ -150,112 +156,124 @@ public class home {
     }
 
 
+    // -----
 
+    // Adds a new home to the homeList of a player
     private static void SetHome(ServerPlayer player, String homeName) throws Exception {
         homeName = homeName.toLowerCase();
-        BlockPos blockPos = new BlockPos(player.getBlockX(), player.getBlockY(), player.getBlockZ());
+        BlockPos blockPos = player.blockPosition();
         ServerLevel world = player.serverLevel();
 
-        Pair<StorageManager.StorageClass, StorageManager.StorageClass.Player> storages = GetPlayerStorage(player.getStringUUID());
-        StorageManager.StorageClass storage = storages.getFirst();
-        StorageManager.StorageClass.Player playerStorage = storages.getSecond();
-
-        boolean homeNotFound = true;
+        // Gets player storage and makes it if it doesn't exist
+        Player playerStorage = StorageManager.STORAGE.addPlayer(player.getStringUUID());
 
         // check for duplicates
-        for (StorageManager.StorageClass.NamedLocation currentHome : playerStorage.Homes) {
-            if (Objects.equals(currentHome.name, homeName)) {
-                homeNotFound = false;
-                break;
-            }
-        }
-
-        if (homeNotFound) {
-            // Create a new NamedLocation
-            StorageManager.StorageClass.NamedLocation homeLocation = new StorageManager.StorageClass.NamedLocation(homeName, blockPos, world.dimension().location().toString());
-
-            playerStorage.Homes.add(homeLocation);
-
-            if (playerStorage.Homes.size() == 1) {
-                playerStorage.DefaultHome = homeName;
-            }
-
-            StorageSaver();
-            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.set", player), true);
-        } else {
+        if (playerStorage.getHome(homeName).isPresent()) {
             player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.exists", player).withStyle(ChatFormatting.RED), true);
+            return;
         }
+
+        // Create a new NamedLocation
+        playerStorage.setHome(homeName, blockPos, world.dimension().location().toString());
+
+        // Set it as the default if there are no other homes
+        if (playerStorage.getHomes().size() == 1) {
+            playerStorage.setDefaultHome(homeName);
+        }
+
+        // Display message that the home as been set
+        player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.set", player), true);
     }
 
+    // Teleports the player to the home. It will go to the defaultHome if homeName is empty
     private static void GoHome(ServerPlayer player, String homeName) throws Exception {
         homeName = homeName.toLowerCase();
-        StorageManager.StorageClass.Player playerStorage = GetPlayerStorage(player.getStringUUID()).getSecond();
 
-        // check if there is a default exists
+        // Gets player storage
+        Optional<Player> optionalPlayerStorage = STORAGE.getPlayer(player.getStringUUID());
+        if (optionalPlayerStorage.isEmpty()) {
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.homeless", player).withStyle(ChatFormatting.AQUA), true);
+            return;
+        }
+
+        Player playerStorage = optionalPlayerStorage.get();
+
+        // if homeName is empty, get the default home
         if (homeName.isEmpty()) {
-            if (playerStorage.DefaultHome.isEmpty()) {
+            // todo! if there is no default home set, maybe give an message saying: no default home set!
+            String defaultHome = playerStorage.getDefaultHome();
+
+            if (defaultHome.isEmpty()) {
                 player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.homeless", player).withStyle(ChatFormatting.AQUA), true);
                 return;
             } else {
-                homeName = playerStorage.DefaultHome;
+                homeName = defaultHome;
             }
         }
 
-        boolean foundWorld = false;
-
-        // find correct home
-        for (StorageManager.StorageClass.NamedLocation currentHome : playerStorage.Homes) {
-            if (Objects.equals(currentHome.name, homeName)) {
-
-                // find correct world
-                for (ServerLevel currentWorld : TeleportCommands.SERVER.getAllLevels()) {
-                    if (Objects.equals(currentWorld.dimension().location().toString(), currentHome.world)) {
-                        foundWorld = true;
-
-                        BlockPos blockPos = new BlockPos(currentHome.x, currentHome.y, currentHome.z);
-                        BlockPos playerBlockPos = new BlockPos(player.getBlockX(), player.getBlockY(), player.getBlockZ());
-
-                        if (!playerBlockPos.equals(blockPos) || player.level() != currentWorld) {
-                            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.go", player), true);
-                            Teleporter(player, currentWorld, new Vec3(currentHome.x + 0.5, currentHome.y, currentHome.z + 0.5));
-                        } else {
-                            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.goSame", player).withStyle(ChatFormatting.AQUA), true);
-                        }
-                        break;
-                    }
-                }
-            }
+        // get the home (if it exists)
+        Optional<NamedLocation> optionalHome = playerStorage.getHome(homeName);
+        if (optionalHome.isEmpty()) {
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.notFound", player).withStyle(ChatFormatting.AQUA), true);
+            return;
         }
 
-        if (!foundWorld) {
-            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.notFound", player).withStyle(ChatFormatting.RED), true);
+        NamedLocation home = optionalHome.get();
+
+        // get the world, otherwise throw an exception
+        Optional<ServerLevel> optionalHomeWorld = home.getWorld();
+        if (optionalHomeWorld.isEmpty()) {
+            // todo! test this exception
+
+            throw new Exception( String.format("Couldn't find a world with the id: %s \nAvailable worlds: %s",
+                    home.getWorldString(), TeleportCommands.SERVER.getAllLevels()));
+        }
+
+        ServerLevel homeWorld = optionalHomeWorld.get();
+        BlockPos teleportBlockPos = home.getBlockPos();
+
+        // check if the player is already at this location (in the same world)
+        if (player.blockPosition().equals(teleportBlockPos) && player.level() == homeWorld) {
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.goSame", player).withStyle(ChatFormatting.AQUA), true);
+
+        } else {
+            // teleport the player!
+            Vec3 teleportPos = new Vec3(teleportBlockPos.getX() + 0.5, teleportBlockPos.getY(), teleportBlockPos.getZ() + 0.5);
+
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.go", player), true);
+            Teleporter(player, homeWorld, teleportPos);
         }
     }
 
     private static void DeleteHome(ServerPlayer player, String homeName) throws Exception {
         homeName = homeName.toLowerCase();
-        Pair<StorageManager.StorageClass, StorageManager.StorageClass.Player> storages = GetPlayerStorage(player.getStringUUID());
-        StorageManager.StorageClass storage = storages.getFirst();
-        StorageManager.StorageClass.Player playerStorage = storages.getSecond();
 
-        boolean deletedHome = false;
-
-        // get correct home
-        for (StorageManager.StorageClass.NamedLocation currentHome : playerStorage.Homes) {
-            if (Objects.equals(currentHome.name, homeName)) {
-                // delete the home
-                playerStorage.Homes.remove(currentHome);
-                StorageSaver();
-
-                deletedHome = true;
-                player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.delete", player), true);
-                break;
-            }
+        // Gets player storage
+        Optional<Player> optionalPlayerStorage = STORAGE.getPlayer(player.getStringUUID());
+        if (optionalPlayerStorage.isEmpty()) {
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.homeless", player).withStyle(ChatFormatting.AQUA), true);
+            return;
         }
 
-        if (!deletedHome) {
-            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.notFound", player).withStyle(ChatFormatting.RED), true);
+        Player playerStorage = optionalPlayerStorage.get();
+
+        // Get the home from the player
+        Optional<NamedLocation> optionalHome = playerStorage.getHome(homeName);
+        if (optionalHome.isEmpty()) {
+            player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.notFound", player).withStyle(ChatFormatting.AQUA), true);
+            return;
         }
+
+        NamedLocation home = optionalHome.get();
+
+        // check if it's the default home, if it is set it to the default value
+        if (playerStorage.getDefaultHome().equals(homeName)) {
+            playerStorage.setDefaultHome("");
+        }
+
+        // delete the home
+        playerStorage.deleteHome(home);
+        player.displayClientMessage(getTranslatedText("commands.teleport_commands.home.delete", player), true);
     }
 
     private static void RenameHome(ServerPlayer player, String homeName, String newHomeName) throws Exception {