+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the Storage classes and functions (I'm doing proper java, yipie)
 - Better error handling for command suggestions
 - Added hover effects for warp and homes text (W.I.P)
+- Throw an exception when the world isn't found (when doing back or home), instead of giving an incorrect notFound error.
 
 ### [v1.2.2]
 - Handled a case where the client (geyser) will return the language as uppercase instead of lowercase.